refactor: move api routes to v1 namespace
