+++ firebase-functions/src/index.ts
@@ -1,7 +1,7 @@
 import { https } from "firebase-functions";
 import admin from "firebase-admin";
-import app from "./api";
+import apiV1 from "./api/v1";
 
 admin.initializeApp();
 
-export const api = https.onRequest(app);
+export const v1 = https.onRequest(apiV1);
